refactor(cef_host): demote to lib-only crate, extract cef_settings module
Splits the former cef_host binary's main.rs into:
- src/cef_settings.rs — build_cef_settings, load_cef_framework,
  acquire_data_root, and the macOS bundle-path helpers
  (in_bundle_helper, in_bundle_framework_dylib)
- src/browser_app.rs — the cef::App impl that injects per-process
  command-line flags
- src/lib.rs — re-exports the items the daemon needs and documents the
  "no CEF refcount types in tokio::spawn" code discipline

Bin-only code (init_tracing, dispatch_helper_process_or_continue,
spawn_control_runtime, main itself) is dropped; ozmux-daemon will own
those in subsequent tasks.

Updates the in-bundle helper lookup from "cef_host Helper.app" to
"ozmux-daemon Helper.app" to match the integrated daemon bundle.

Part of Plan 3 (cef_host -> lib + ozmux-daemon merge), Task 3.
