+++ daemon/cef_host/src/lib.rs
@@ -1,9 +1,20 @@
-//! Library entry for the ozmux cef_host crate. Exposes internal modules so
-//! tests can exercise them. The actual binary (`bin/cef_host`) and helper
-//! (`bin/cef_helper`) use these via the crate path.
+//! Library entry for the ozmux cef_host crate. Exposes the CEF host machinery
+//! (CEF settings, app implementation, browser pool, UDS control plane,
+//! handlers, IPC) consumed by the `ozmux-daemon` binary and the `cef_helper`
+//! subprocess.
+//!
+//! # Code discipline
+//!
+//! CEF refcount-bearing types (`Browser`, `RequestContext`, `OzmuxRenderHandler`,
+//! `PoolHandle`, etc.) must NEVER be captured into a `tokio::spawn`ed future.
+//! Their `Drop` may release CEF objects, which must happen on the CEF UI thread.
+//! Use `post_command::post` to schedule mutation (including teardown) on the UI
+//! thread instead.
 
 use cef::ImplCommandLine;
 
+pub mod browser_app;
+pub mod cef_settings;
 pub mod control;
 pub mod cookies;
 pub mod handlers;
@@ -12,6 +23,11 @@ pub mod pool;
 pub mod post_command;
 pub mod profile;
 pub mod shm_writer;
+
+pub use browser_app::BrowserApp;
+pub use cef_settings::{acquire_data_root, build_cef_settings, load_cef_framework};
+#[cfg(target_os = "macos")]
+pub use cef_settings::{in_bundle_framework_dylib, in_bundle_helper};
 
 /// Appends a flag-only switch to a CEF command line. Compresses the
 /// boilerplate of constructing a `cef::CefString` per call.
